WIP - Contamination Zeus Module
What
=================
So far successfully added the module to Zeus, and it creates a dialog.
All that needs to be done now is for the Module to use the data from the
Dialog, as well as some of the settings normally applied through the
3DEN modules, to be applied via CBA settings or hardcoded instead, and
it would be finished.

Issue ID: N/A