WIP - Contamination Zeus Module
What
=================
Contamination module itself works, and contains no more errors. However,
the module fails to create the contamination zone itself. This will be
looked at again in the future. For now, it is not game breaking,
therefore can be left in here. Usually a separate development branch for
this feature would be created, but I failed to do so before pushing the
first WIP commit, so for now it can stay.

Issue ID: #3