Added/Changed - Sling Mask: Ability to Swap facewear for slung Gas Mask
What
=================
Added: Ability to "swap" the facewear (glasses, goggles, etc.) a user is wearing with the slung
Gas Mask. This would mean no need to access inventory when needing to
unsling your Gas Mask! The Gas Mask will be unslung and your goggles
that you may have previously been wearing is slung (if allowed) or added
to your inventory.
BEWARE: If there is not enough space for the facewear item in your
inventory, then the facewear item will be dropped onto the floor!
Changed: naming of any functions defined in this mod with the
appropriate prefixes and added appropriate documentation, so that
functions and such from the original "GRAD Sling Helmet" mod are
distinguishable from those defined in this mod.

Issue ID: #13