Added/Changed - Build script change detection + obsolete pbo removal
What
=================
Added: Ability for the build script to detect whether or not the
contents of an addon has actually changed before rebuilding it. If not,
then leave it be. If yes, then update the built addon PBO.
Added: Ability for the build script to delete any PBOs from the
`buildDir\addons` folder that do not exist as addons in the project anymore.
Changed: Made the mod prefix a variable for when the script is used for
other projects.

NOTE: PBO signkey files aren't synced with these changes yet.

Issue ID: #20