Changed/Fixed - pre/postInit names + ACE Default loadouts not showing
What
=================
Changed: the names of the pre/postInit files to be normalised.
Fixed: Default loadouts not showing. For some reason, since the addition
of an extra `Ares_fnc_RegisterCustomModule` the mission didn't seem to
be loading the default loadouts in time? Or maybe too early? Either way,
by moving that code to postInit I've fixed the problem.

Yes. I did a Ford, used Slav magic, and am satisfied to leave it as is,
as long as it's working.

*shrug*

Why
=================
Because Ford's Slav magic is awesome.

Issue ID: #51