Added - Zeus Module: Add Filtered Arsenal ability to spawn object
What
=================
Added: ability for the existing Zeus arsenal to now spawn an object from
a dropdown list if module is not placed on an object. Bear in mind that,
it the selection list is kept blank, the module will fall back to
allowing a selection of multiple objects on which to apply the arsenal.

End result: A completely flexible module.

Issue ID: #27