Fixed - SMA error showing when BNorway vehicles used
What
=================
Replaced: SMA HK416 with RHS's version of the HK416 D10.
Added: Ammo for the 3CB vehicles, and inventories for some more vehicles
that needed it or had MX's. Left the inventories with M4s alone though.

Why
=================
Because we used to have the SMA mod in our modpack, but not anymore, so
the game didn't much like that...

Issue ID: #62 #71