Added - Screenshot Mode by Gruppe Adler (GRAD)
Why
=================
To allow our players to have a clean screen for screenshots or
recordings. This removes UI elements such as the top left and right
boxes showing ammo count, vehicle stats, etc. as well as the DUI squad
radar on the bottom.

This does NOT hide the side chat nor any hint that pops up, sadly...

Issue ID: N/A