Added - Swap feature except for helmets - TEMPORARY
What
=================
Added the swap feature, where you swap an item you're wearing for the
item that you have slung, to be used for helmets in addition to the
pre-existing swap mask feature. This is only temporary, until
https://github.com/DerZade/grad_slinghelmet/pull/23 gets merged.

Why
=================
Because I'm an impatient bastard who likes his cool stuff!

Issue ID: N/A