Fixed - Full Heal: caused duplicates and server crashes/desyncs
Why
=================
We're not entirely sure why the server desync and crashes were
happening... only that they were most likely caused by the code related
to the full heal feature. So, it has been fixed to optimise it in terms
of the network overhead, and any errors that were being thrown.

Issue ID: #88