Added - Default Loadouts only show those relevant to map climate
What
=================
Added: the possibility to only show default loadouts that fit the
climate of any given map in the modpack. Simply select what climate
should any of the given maps be considered to have, and only the default
loadouts for that climate will be visible.

Why
=================
To have less crap in the default loadouts screen, and also to avoid
confusion as to which uniform is considered appropriate for any given
map.

Issue ID: #68