Added/Fixed - Player Killed notification for Zeus & respawn counter
What
=================
Added: Player killed notification for a Zeus whilst he's in the Zeus
interface. This does not show when anyone is not in the Zeus interface.
Fixed: Respawn timer in CBA Settings wasn't being applied before, now it
is. This means the option is no longer dependent on someone setting up
the time delay in the mission itself. The respawn system does still have
to be applied, however.

Issue ID: https://github.com/Saborknight/2bnb-framework/issues/10