Fixed - Filtered Arsenal filter not consistently working
What
=================
Fixed: I THINK, where the Arsenal filter isn't always being applied
correctly. This was most notable on someone JIPing. This lead me to
believe it was because of some kind or re-initing of the Arsenal. So
here's me fixing the fact that `bnb_f_core_fnc_arsenal` was getting
executed again and again everytime someone JIPed.

Issue ID: #94