Fixed - Respawn all command resetting to the server's respawn timer
What
=================
Fixed: how the server respawn timer kept on being used when the module
was used. Now the server's respawn timer value will only be used if it's
actually changed, therefore we assume that it was changed because the
user actually wants to use a new respawn timer value.

Issue ID: N/A