Added - Civilian interaction using ACE gestures
What
=================
Added the ability to interact with civilians to either have them wave
back, stop, go away or follow a player. Stop, go away and follow depend
on whether or not the player and the civilians have weapons. If the player
does and the civilians do not, then the civilians will comply. If the
player does not, and the civilians do, then the civilians are less likely
to comply.

Issue ID: #134