Added - Made 2BNB Beret carryable in uniform and vest, and weightless
What
=================
Changed: Mass and allowedSlots values of the 2BNB Beret to make is
weightless and allow it to be stored in uniform, vest, backpack and in
the headgear slot of the character.
Added: To all uniforms of the default loadouts. I didn't include them in
the faction unit loadouts yet because of how it's fucked right now, and
it's a decorative change anyway.

Why
=================
So that it could be included in the loadout of every role, regardless
if they have space in their loadouts or not.

Issue ID: #128