Fixed - I hope: Player death duration not showing in death list
What
=================
Fixed: The duration for which the player has been dead when the list of
the currently dead players (who are assumed sitting in spectator at that
time) is brought up, as per #92.

I hope this fixes it, since I can't test it's effect without testing on
multiplayer with others.

Issue ID: #92