Added - Norwegian Task Force assets & Armoured vic loadouts into macro
What
=================
Added: NTF assets which were imported with permission. Included without
modification are the Bell 412 and the C-130J. Included with modification
are the M113, the CV90, and the Scania trucks. The modifications
included heavy colour correction to fit the existing RHS Humvees and
M113s, fixes and such to the textures that had incorrect textures on
some points of the models.
Added: Armoured Vehicle loadouts macro to standardise the inventories of
the vehicles in the BNorway faction.
Removed: FMTV truck since the Scanias would be replacing it as general
purpose heavy transportation.

Issue ID: #151