Changed - Respawn timer mission config setting imports to CBA setting
What
=================
Changed: the respawn timer to, instead of making an if statement to
check which timer to use (vanilla mission config timer, or the one set
in CBA settings) every time, it now just imports the vanilla config
setting into the CBA setting IF the CBA setting has nothing set in the
mission CBA setting beforehand.
This means the override variable is no longer necessary.

Issue ID: #159