Fixed - Zeus Tools: Issues with timer UI and respawn timer being set
What
=================
Fixed: The Timer UI that is opened in the Zeus ACE interact, now it will
display the currently in use respawn time when opened.
Fixed: Respawn timer applying the new respawn time as being that which
just changed, if server was just changed, it would apply that, if
mission, then that, regardless of the order of priority (AKA whether or
not the server is supposed to override mission). So now it will only
ever apply the prioritised value, AKA bnb_e_respawn_timer. Since the
Respawn Timer UI relies on this CBA Setting "onChange" code to apply its
changes also, it is also affected.

Issue ID: #161