Added/Fixed - Aviation inventories & Bad Vehicle Type for some vics
What
=================
Added: Inventories for air vehicles. This inventory includes a basic
survival backpack that contains basic medical equipment, as well as a
surgical kit and other tidbits.
Fixed: Bad Vehicle Type error popping up whenever a vehicle with the
toolkit backpack was present. Turns out the backpack needs to be defined
with `backpack = ` instead of `name = `.

Issue ID: N/A