Added - VBED: Only disarm with defuse tool, and EOD (if set on server)
What
=================
Added: A restriction that means only those that have defusal tools and
EOD trait is able to disarm an explosive in a vehicle. This takes the
ACE setting "Require Specialist" as the setting that defines whether or
not the person needs to be EOD. Basically it's the same functionality
that an ACE Explosive has in terms of checking whether or not a player
should be able to disarm an explosive.

Issue ID: #135