Added - Russian Winter Retextures: fixed some of the mod's units
What
=================
Added: The crew unit that was implemented in the Russian Winter
Retextures mod to the `units[]` array. It was missing there in the
original mod, which then broke all vehicles with that unit as their
crew, because when Arma 3 can't find the crew for a vehicle, then
apparently that means the whole damn vehicle shouldn't show up in
Zeus......... Logic? Pfff, fuck that.

Issue ID: N/A